[Fix #25] Correctly render offline, non-BNET friends
When rendering the row for an offline, non-BNET friend, the
variable `nameText` was being set to the value of the local
variable `name` variable. `name` was never defined and therefore
had a value of `nil`.

Apparently, setting the text to `nil` causes that entire line to
appear blank (no textures, no nothing). We should get the friend's
name from `info.name` instead.